jni: minisdl: individal module

+++ ijkmediaplayer/jni/ijkplayer/ijkplayer_jni.c
@@ -49,7 +49,7 @@ static IjkMediaPlayer *get_media_player(JNIEnv* env, jobject thiz)
     // FIXME: lock ref count
     pthread_mutex_lock(&g_clazz.mutex);
 
-    IjkMediaPlayer *mp = (IjkMediaPlayer *) (*env)->GetLongField(env, thiz, g_clazz.mNativeMediaPlayer);
+    IjkMediaPlayer *mp = (IjkMediaPlayer *) (intptr_t) (*env)->GetLongField(env, thiz, g_clazz.mNativeMediaPlayer);
     if (mp) {
         ijkmp_inc_ref(mp);
     }
